+++ src/routes/api/v1/health.router.ts
@@ -2,7 +2,6 @@ import { Router } from 'express';
 import {
   getHealth,
   getReadiness,
-  firstLogin,
 } from '../../../controllers/health.controller';
 
 const router = Router();
@@ -10,7 +9,5 @@ const router = Router();
 router.get('/liveness', getHealth);
 
 router.get('/readiness', getReadiness);
-
-router.post('/firstLogin', firstLogin);
 
 export default router;
