+++ src/routes/api/v1/usuario.route.ts
@@ -1,0 +1,10 @@
+import { Router } from 'express';
+import {
+    firstLogin,
+} from '../../../controllers/usuario.controller';
+
+const router = Router();
+
+router.post('/firstLogin', firstLogin);
+
+export default router;
